Added ms2_stopwatch.F90 to implement a timer functionality and included the new file in Makefile and Makefile.mult Now Subroutine TSimulation_Run (ms2_simulation.F90) uses TStopwatch to measure time.
Changed FileReadParameter (ms2_global.F90) and added optional parameters "rewind_before" and "defaultvalue".
With rewind_before =.true. the file will be rewinded (reading position will be set to beginning) before reading.
To some extend, the order of unique parameters within a file is more free using this hack.
Setting defaultvalue="value" the subroutine will set IObuffer to "value" and return instead of terminating the program,
if the parameter is not found. There should be a defaultvalue given, if possible/applicable and old files might still be usable
when new parameters with default values are introduced.


git-svn-id: https://scm.projects.hlrs.de/authscm/sleyh/svn/ms2/trunk@24 a2be33b2-7eaf-4dce-bf61-b02fde6c56c2