BUGFIX
ms2_interaction.F90:637f
MayerFFunction* and IntFFunction* are only allocated for SimulationType .eq. SecondVirialCoeff
Deallocation should also be done only for this case or alternatively only if pointers are associated (the latter is realized, line 757f).
BUT POINTERS HAVE TO BE NULLIFIED IF THERE'S NO ALLOCATION THEN!

ms2_simulation.F90:1083
maximum line length for a FORTRAN program is 132 columns (btw: also with comments)


git-svn-id: https://scm.projects.hlrs.de/authscm/sleyh/svn/ms2/trunk@145 a2be33b2-7eaf-4dce-bf61-b02fde6c56c2